+++ tests/LiftGenericsSpec.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE CPP #-}
-{-# LANGUAGE DeriveDataTypeable #-}
 {-# LANGUAGE TemplateHaskell #-}
 
 {-|
@@ -15,16 +14,13 @@ module LiftGenericsSpec (main, spec) where
 import Instances.TH.Lift ()
 import Language.Haskell.TH.Syntax hiding (newName)
 import Language.Haskell.TH.Syntax.Compat
-#if MIN_VERSION_base(4,5,0)
 import Language.Haskell.TH.Lift.Generics (genericLift)
-# if MIN_VERSION_template_haskell(2,16,0)
+#if MIN_VERSION_template_haskell(2,16,0)
 import Language.Haskell.TH.Lift.Generics (genericLiftTypedCompat)
-# endif
 #endif
 import Test.Hspec
 import Types
 import Control.Exception (Exception, throw, evaluate)
-import Data.Typeable (Typeable)
 
 main :: IO ()
 main = hspec spec
@@ -32,7 +28,7 @@ main = hspec spec
 description :: String
 description = "should equal its lifted counterpart"
 
-data Exc = Exc deriving (Show, Typeable, Eq)
+data Exc = Exc deriving (Show, Eq)
 instance Exception Exc
 
 spec :: Spec
@@ -59,7 +55,6 @@ spec = parallel $ do
         describe "Ap" $
             it description $
                 z `shouldBe` $(lift z)
-#if MIN_VERSION_base(4,5,0)
         -- We use Tree to check that things work for types imported
         -- from external packages without special distinguished names.
         -- It proved rather tricky to find a good choice. This one's
@@ -68,30 +63,7 @@ spec = parallel $ do
         describe "Tree" $
             it description $
                 w `shouldBe` $(genericLift w)
-        describe "Unit'" $ do
-            it description $ do
-                Unit' `shouldBe` $(lift Unit')
-                ConE 'Unit' `shouldBe` runPureQ (liftQuote Unit')
-            it "should throw an exception on undefined" $
-                evaluate (runPureQ $ liftQuote (throw Exc :: Unit')) `shouldThrow` (== Exc)
-        describe "Product'" $
-            it description $
-                p' `shouldBe` $(lift p')
-        describe "Sum'" $
-            it description $
-                s' `shouldBe` $(lift s')
-        describe "Unboxed'" $
-            it description $
-                u' `shouldBe` $(lift u')
-        describe "Const'" $
-            it description $
-                c' `shouldBe` $(lift c')
-# if __GLASGOW_HASKELL__ >= 708
-        describe "Ap'" $
-            it description $
-                z' `shouldBe` $(lift z')
-#endif
-# if MIN_VERSION_template_haskell(2,16,0)
+#if MIN_VERSION_template_haskell(2,16,0)
     describe "genericLiftTyped" $ do
         describe "Unit" $
             it description $ do
@@ -115,5 +87,4 @@ spec = parallel $ do
         describe "Tree" $
             it description $
                 w `shouldBe` $$(genericLiftTypedCompat w)
-# endif
 #endif